requirements.txt
scp
paramiko
youtube-dl
python-telegram-bot
